requirements.txt
Flask==2.3.2
Flask-Login==0.6.2
python-dotenv==1.0.0
gunicorn==20.1.0